+++ torch/csrc/generic/serialization.cpp
@@ -2,25 +2,25 @@
 #define TH_GENERIC_FILE "generic/serialization.cpp"
 #else
 
-#define SYSCHECK(call) { ssize_t __result = call; if (__result < 0) throw std::system_error(__result, std::system_category()); }
+#define SYSCHECK(call) { ssize_t __result = call; if (__result < 0) throw std::system_error((int) __result, std::system_category()); }
 
 void THPTensor_(writeMetadataRaw)(THTensor *self, int fd)
 {
-  SYSCHECK(write(fd, &self->nDimension, sizeof(long)));
-  SYSCHECK(write(fd, self->size, sizeof(long) * self->nDimension));
-  SYSCHECK(write(fd, self->stride, sizeof(long) * self->nDimension));
-  SYSCHECK(write(fd, &self->storageOffset, sizeof(long)));
+  SYSCHECK(write(fd, &self->nDimension, sizeof(int64_t)));
+  SYSCHECK(write(fd, self->size, sizeof(int64_t) * self->nDimension));
+  SYSCHECK(write(fd, self->stride, sizeof(int64_t) * self->nDimension));
+  SYSCHECK(write(fd, &self->storageOffset, sizeof(int64_t)));
 }
 
 THTensor * THPTensor_(newWithMetadataFileRaw)(int fd, THStorage *storage)
 {
   THTensorPtr tensor = THTensor_(new)(LIBRARY_STATE_NOARGS);
-  SYSCHECK(read(fd, &tensor->nDimension, sizeof(long)));
-  tensor->size = (long*)THAlloc(tensor->nDimension * sizeof(long));
-  tensor->stride = (long*)THAlloc(tensor->nDimension * sizeof(long));
-  SYSCHECK(read(fd, tensor->size, sizeof(long) * tensor->nDimension));
-  SYSCHECK(read(fd, tensor->stride, sizeof(long) * tensor->nDimension));
-  SYSCHECK(read(fd, &tensor->storageOffset, sizeof(long)));
+  SYSCHECK(read(fd, &tensor->nDimension, sizeof(int64_t)));
+  tensor->size = (int64_t*)THAlloc(tensor->nDimension * sizeof(int64_t));
+  tensor->stride = (int64_t*)THAlloc(tensor->nDimension * sizeof(int64_t));
+  SYSCHECK(read(fd, tensor->size, sizeof(int64_t) * tensor->nDimension));
+  SYSCHECK(read(fd, tensor->stride, sizeof(int64_t) * tensor->nDimension));
+  SYSCHECK(read(fd, &tensor->storageOffset, sizeof(int64_t)));
   THStorage_(retain)(LIBRARY_STATE storage);
   tensor->storage = storage;
   return tensor.release();
@@ -36,14 +36,14 @@ void THPStorage_(writeFileRaw)(THStorage *self, int fd)
   data = (real*)cpu_data.get();
   THCudaCheck(cudaMemcpy(data, self->data, self->size * sizeof(real), cudaMemcpyDeviceToHost));
 #endif
-  SYSCHECK(write(fd, &self->size, sizeof(long)));
+  SYSCHECK(write(fd, &self->size, sizeof(int64_t)));
   // fast track for bytes and little endian
   if (sizeof(real) == 1 || THP_nativeByteOrder() == THPByteOrder::THP_LITTLE_ENDIAN) {
     SYSCHECK(write(fd, data, sizeof(real) * self->size));
   } else {
-    long buffer_size = std::min(self->size, (long)5000);
+    int64_t buffer_size = std::min(self->size, (int64_t)5000);
     std::unique_ptr<uint8_t[]> le_buffer(new uint8_t[buffer_size * sizeof(real)]);
-    for (long i = 0; i < self->size; i += buffer_size) {
+    for (int64_t i = 0; i < self->size; i += buffer_size) {
       size_t to_convert = std::min(self->size - i, buffer_size);
       if (sizeof(real) == 2) {
         THP_encodeInt16Buffer((uint8_t*)le_buffer.get(),
@@ -69,8 +69,8 @@ void THPStorage_(writeFileRaw)(THStorage *self, int fd)
 THStorage * THPStorage_(readFileRaw)(int fd)
 {
   real *data;
-  long size;
-  SYSCHECK(read(fd, &size, sizeof(long)));
+  int64_t size;
+  SYSCHECK(read(fd, &size, sizeof(int64_t)));
   THStoragePtr storage = THStorage_(newWithSize)(LIBRARY_STATE size);
 
 #ifndef THC_GENERIC_FILE
@@ -84,9 +84,9 @@ THStorage * THPStorage_(readFileRaw)(int fd)
   if (sizeof(real) == 1 || THP_nativeByteOrder() == THPByteOrder::THP_LITTLE_ENDIAN) {
     SYSCHECK(read(fd, data, sizeof(real) * storage->size));
   } else {
-    long buffer_size = std::min(size, (long)5000);
+    int64_t buffer_size = std::min(size, (int64_t)5000);
     std::unique_ptr<uint8_t[]> le_buffer(new uint8_t[buffer_size * sizeof(real)]);
-    for (long i = 0; i < size; i += buffer_size) {
+    for (int64_t i = 0; i < size; i += buffer_size) {
       size_t to_convert = std::min(size - i, buffer_size);
       SYSCHECK(read(fd, le_buffer.get(), sizeof(real) * to_convert));
       if (sizeof(real) == 2) {
