+++ torch/lib/THC/generic/THCTensorMath.cu
@@ -79,7 +79,7 @@ void THCTensor_(catArray)(THCState *state, THCTensor *result,
 {
   THLongStorage *size;
   int i, j, cohortMax;
-  long offset;
+  int64_t offset;
   bool hasEmptyInput = false;
 
   // Even in the case where dimension is negative (i.e. when we want
@@ -112,7 +112,7 @@ void THCTensor_(catArray)(THCState *state, THCTensor *result,
   for(i = 0; i < maxDim; i++)
   {
     // dimSize is either the size of the dim if it exists, either 1 if #dim > 0, otherwise 0
-    long dimSize = i < THCTensor_(nDimension)(state, inputs[0])
+    int64_t dimSize = i < THCTensor_(nDimension)(state, inputs[0])
                        ? THCTensor_(size)(state, inputs[0], i)
                        : THMin(THCTensor_(nDimension)(state, inputs[0]), 1);
     if (i == ldimension)
@@ -130,7 +130,7 @@ void THCTensor_(catArray)(THCState *state, THCTensor *result,
     {
       for (j = 1; j < numInputs; j++)
       {
-        long sz = i < THCTensor_(nDimension)(state, inputs[j])
+        int64_t sz = i < THCTensor_(nDimension)(state, inputs[j])
                       ? THCTensor_(size)(state, inputs[j], i)
                       : THMin(THCTensor_(nDimension)(state, inputs[j]), 1);
 
@@ -210,7 +210,7 @@ void THCTensor_(catArray)(THCState *state, THCTensor *result,
     for (i = 0; i < numInputs; i += CAT_ARRAY_BATCH_SIZE) {
       cohortMax = 0;
       for (j = 0; j < CAT_ARRAY_BATCH_SIZE && (i+j) < numInputs; ++j) {
-        long dimSize = ldimension < THCTensor_(nDimension)(state, inputs[i+j])
+        int64_t dimSize = ldimension < THCTensor_(nDimension)(state, inputs[i+j])
           ? THCTensor_(size)(state, inputs[i+j], ldimension)
           : 1;
 
@@ -267,7 +267,7 @@ void THCTensor_(catArray)(THCState *state, THCTensor *result,
       // No reason to copy when input is empty
       if (!THCTensor_(nDimension)(state, inputs[j])) continue;
 
-      long dimSize = ldimension < THCTensor_(nDimension)(state, inputs[j])
+      int64_t dimSize = ldimension < THCTensor_(nDimension)(state, inputs[j])
                ? THCTensor_(size)(state, inputs[j], ldimension)
                : 1;
 
@@ -293,16 +293,16 @@ void THCTensor_(nonzero)(THCState* state, THCudaLongTensor *tensor,
   thrust::device_ptr<real> self_data(THCTensor_(data)(state, self));
 
   int num_dim = THCTensor_(nDimension)(state, self);
-  long N = THCTensor_(nElement)(state, self);
+  int64_t N = THCTensor_(nElement)(state, self);
 
   THCudaLongTensor_resize2d(state, tensor, N, num_dim);
   tensor = THCudaLongTensor_newContiguous(state, tensor);
-  thrust::device_ptr<long> tensor_data(THCudaLongTensor_data(state, tensor));
+  thrust::device_ptr<int64_t> tensor_data(THCudaLongTensor_data(state, tensor));
 
-  thrust::counting_iterator<long> idxfirst(0);
-  thrust::counting_iterator<long> idxlast = idxfirst + N;
+  thrust::counting_iterator<int64_t> idxfirst(0);
+  thrust::counting_iterator<int64_t> idxlast = idxfirst + N;
 
-  typedef thrust::device_ptr<long> Iter;
+  typedef thrust::device_ptr<int64_t> Iter;
   strided_range<Iter> strided_tensor(tensor_data,
                                      tensor_data+N*num_dim, num_dim);
 
@@ -321,9 +321,9 @@ void THCTensor_(nonzero)(THCState* state, THCudaLongTensor *tensor,
     NonZeroOp<real>()
   );
 
-  long num_nonzeros = thrust::distance(strided_tensor.begin(), dend);
+  int64_t num_nonzeros = thrust::distance(strided_tensor.begin(), dend);
 
-  long div = 1;
+  int64_t div = 1;
   for (int dim = num_dim-1; dim >= 0; dim--) {
     strided_range<Iter> stride_dim(tensor_data+dim,
                                    tensor_data+N*num_dim, num_dim);
@@ -347,33 +347,33 @@ void THCTensor_(nonzero)(THCState* state, THCudaLongTensor *tensor,
   THCudaCheck(cudaGetLastError());
 }
 
-void THCTensor_(diag)(THCState *state, THCTensor *self_, THCTensor *src_, long k){
+void THCTensor_(diag)(THCState *state, THCTensor *self_, THCTensor *src_, int64_t k){
   THAssert(THCTensor_(checkGPU)(state, 2, self_, src_));
   int nDimension = THCTensor_(nDimension)(state, src_);
   THArgCheck((nDimension == 2) || (nDimension == 1), 1, "expected a matrix or a vector");
   if (nDimension == 2) {
-    long stride0 = THCTensor_(stride)(state, src_, 0);
-    long stride1 = THCTensor_(stride)(state, src_, 1);
-    long size0 = THCTensor_(size)(state, src_, 0);
-    long size1 = THCTensor_(size)(state, src_, 1);
-    long size = (k > 0) ? min((long long)size0, (long long)size1 - k) : min((long long)size0 + k, (long long)size1);
+    int64_t stride0 = THCTensor_(stride)(state, src_, 0);
+    int64_t stride1 = THCTensor_(stride)(state, src_, 1);
+    int64_t size0 = THCTensor_(size)(state, src_, 0);
+    int64_t size1 = THCTensor_(size)(state, src_, 1);
+    int64_t size = (k > 0) ? min((int64_t)size0, (int64_t)size1 - k) : min((int64_t)size0 + k, (int64_t)size1);
     THCTensor_(resize1d)(state, self_, size);
-    long strideSelf = THCTensor_(stride)(state, self_, 0);
-    const dim3 threads(min((long long)THCState_getCurrentDeviceProperties(state)->maxThreadsPerBlock, (long long)size));
-    dim3 grid(min((long long)1024, (long long)THCCeilDiv(size, (long)threads.x)));
-    long start = (k >= 0 ? k * stride1 : -k * stride0);
+    int64_t strideSelf = THCTensor_(stride)(state, self_, 0);
+    const dim3 threads(min((int64_t)THCState_getCurrentDeviceProperties(state)->maxThreadsPerBlock, (int64_t)size));
+    dim3 grid(min((int64_t)1024, (int64_t)THCCeilDiv(size, (int64_t)threads.x)));
+    int64_t start = (k >= 0 ? k * stride1 : -k * stride0);
     THCTensor_copyFromDiagonal<real><<<grid, threads, 0, THCState_getCurrentStream(state)>>>
     (THCTensor_(data)(state, self_), THCTensor_(data)(state, src_), start, size, stride0 + stride1, strideSelf);
   } else {
     ptrdiff_t totalElements = THCTensor_(nElement)(state, src_);
     ptrdiff_t size = (k > 0) ? totalElements + k : totalElements - k;
-    long strideSrc = THCTensor_(stride)(state, src_, 0);
+    int64_t strideSrc = THCTensor_(stride)(state, src_, 0);
     THCTensor_(resize2d)(state, self_, size, size);
     THCTensor_(zero)(state, self_);
-    long stride0 = THCTensor_(stride)(state, self_, 0);
-    long stride1 = THCTensor_(stride)(state, self_, 1);
-    const dim3 threads(min((long long)THCState_getCurrentDeviceProperties(state)->maxThreadsPerBlock, (long long)size));
-    dim3 grid(min((long long)1024, (long long)THCCeilDiv(size, (ptrdiff_t)threads.x)));
+    int64_t stride0 = THCTensor_(stride)(state, self_, 0);
+    int64_t stride1 = THCTensor_(stride)(state, self_, 1);
+    const dim3 threads(min((int64_t)THCState_getCurrentDeviceProperties(state)->maxThreadsPerBlock, (int64_t)size));
+    dim3 grid(min((int64_t)1024, (int64_t)THCCeilDiv(size, (ptrdiff_t)threads.x)));
     ptrdiff_t start = (k >= 0 ? k * stride1 : -k * stride0);
     THCTensor_copyToDiagonal<real><<<grid, threads, 0, THCState_getCurrentStream(state)>>>
     (THCTensor_(data)(state, self_), THCTensor_(data)(state, src_), start, totalElements, stride0 + stride1, strideSrc);
