+++ torch/lib/THC/generic/THCTensorMathScan.cu
@@ -4,13 +4,13 @@
 
 template<class BinaryOp>
 __host__ void THCTensor_(scanOuterDim)(THCState *state, THCTensor *tgt,
-                                       THCTensor *src, long dimension,
+                                       THCTensor *src, int dimension,
                                        real init, BinaryOp binary_op)
 {
   unsigned ndim = THCTensor_(nDimension)(state, src);
   // Treat all outer dimensions (i.e. dim < dimension) as one.
   unsigned num_orows = 1;
-  for (long dim = 0; dim < dimension; dim++) {
+  for (int dim = 0; dim < dimension; dim++) {
     num_orows *= THCTensor_(size)(state, src, dim);
   }
   unsigned row_size = THCTensor_(size)(state, src, dimension);
@@ -55,7 +55,7 @@ __host__ void THCTensor_(scanInnermostDim)(THCState *state, THCTensor *tgt,
 
 template<class BinaryFunction>
 void THCTensor_(scanDim)(THCState *state, THCTensor *self_, THCTensor *src,
-                         long dimension, real init, BinaryFunction binary_op)
+                         int dimension, real init, BinaryFunction binary_op)
 {
   THCTensor_(resizeAs)(state, self_, src);
 
@@ -72,14 +72,14 @@ void THCTensor_(scanDim)(THCState *state, THCTensor *self_, THCTensor *src,
   THCTensor_(freeCopyTo)(state, self, self_);
 }
 
-void THCTensor_(cumsum)(THCState *state, THCTensor *self, THCTensor *src, long dimension)
+void THCTensor_(cumsum)(THCState *state, THCTensor *self, THCTensor *src, int dimension)
 {
   THAssert(THCTensor_(checkGPU)(state, 2, self, src));
   return THCTensor_(scanDim)(state, self, src, dimension,
                              ScalarConvert<float, real>::to(0.0), AddOp<real>());
 }
 
-void THCTensor_(cumprod)(THCState *state, THCTensor *self, THCTensor *src, long dimension)
+void THCTensor_(cumprod)(THCState *state, THCTensor *self, THCTensor *src, int dimension)
 {
   THAssert(THCTensor_(checkGPU)(state, 2, self, src));
   return THCTensor_(scanDim)(state, self, src, dimension,
